Seperate node modal components and condense post-message functions
