+++ webview-react-app/src/index.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import { VsCodeApiProvider } from './VsCodeApiContext'; // Make sure you import it from the correct file
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "./App";
+import { VsCodeApiProvider } from "./VsCodeApiContext"; // Make sure you import it from the correct file
 
 //wrap app in VsCodeApiProvider component
 ReactDOM.render(
@@ -10,5 +10,5 @@ ReactDOM.render(
       <App />
     </React.StrictMode>
   </VsCodeApiProvider>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
